requirements.txt
flask
psycopg2
flask_sqlalchemy
python-dotenv
sentence-transformers
requests
feedparser
flask_sqlalchemy